Move all required items for Level to trview.app
This means I can start writing tests for it.
