+++ trview.app/Elements/Entity.h
