+++ trview.app/Elements/Room.h
