+++ trview.app/Elements/StaticMesh.cpp
@@ -1,4 +1,3 @@
-#include "stdafx.h"
 #include "StaticMesh.h"
 #include <trview.app/Geometry/Mesh.h>
 #include <trview.app/Geometry/TransparencyBuffer.h>
