+++ trview.app/Elements/StaticMesh.h
