+++ trview.app/Graphics/IMeshStorage.cpp
@@ -1,4 +1,3 @@
-#include "stdafx.h"
 #include "IMeshStorage.h"
 
 namespace trview
