+++ trview.app/Graphics/IMeshStorage.h
