+++ trview.app/Graphics/MeshStorage.cpp
@@ -1,4 +1,3 @@
-#include "stdafx.h"
 #include "MeshStorage.h"
 
 namespace trview
