+++ trview.app/Graphics/MeshStorage.h
