+++ trview.app/trview.app.vcxproj
@@ -25,8 +25,12 @@
     <ClCompile Include="Camera\FreeCamera.cpp" />
     <ClCompile Include="Camera\ICamera.cpp" />
     <ClCompile Include="Camera\OrbitCamera.cpp" />
+    <ClCompile Include="Elements\Entity.cpp" />
     <ClCompile Include="Elements\Item.cpp" />
+    <ClCompile Include="Elements\Level.cpp" />
+    <ClCompile Include="Elements\Room.cpp" />
     <ClCompile Include="Elements\Sector.cpp" />
+    <ClCompile Include="Elements\StaticMesh.cpp" />
     <ClCompile Include="Elements\Trigger.cpp" />
     <ClCompile Include="Geometry\IRenderable.cpp" />
     <ClCompile Include="Geometry\Mesh.cpp" />
@@ -35,8 +39,10 @@
     <ClCompile Include="Geometry\TransparencyBuffer.cpp" />
     <ClCompile Include="Geometry\TransparentTriangle.cpp" />
     <ClCompile Include="Graphics\ILevelTextureStorage.cpp" />
+    <ClCompile Include="Graphics\IMeshStorage.cpp" />
     <ClCompile Include="Graphics\ITextureStorage.cpp" />
     <ClCompile Include="Graphics\LevelTextureStorage.cpp" />
+    <ClCompile Include="Graphics\MeshStorage.cpp" />
     <ClCompile Include="Graphics\SectorHighlight.cpp" />
     <ClCompile Include="Graphics\SelectionRenderer.cpp" />
     <ClCompile Include="Graphics\TextureStorage.cpp" />
@@ -79,9 +85,13 @@
     <ClInclude Include="Camera\FreeCamera.h" />
     <ClInclude Include="Camera\ICamera.h" />
     <ClInclude Include="Camera\OrbitCamera.h" />
+    <ClInclude Include="Elements\Entity.h" />
     <ClInclude Include="Elements\Item.h" />
+    <ClInclude Include="Elements\Level.h" />
+    <ClInclude Include="Elements\Room.h" />
     <ClInclude Include="Elements\RoomInfo.h" />
     <ClInclude Include="Elements\Sector.h" />
+    <ClInclude Include="Elements\StaticMesh.h" />
     <ClInclude Include="Elements\Trigger.h" />
     <ClInclude Include="Elements\Types.h" />
     <ClInclude Include="Geometry\IRenderable.h" />
@@ -94,8 +104,10 @@
     <ClInclude Include="Geometry\TransparentTriangle.h" />
     <ClInclude Include="Geometry\Triangle.h" />
     <ClInclude Include="Graphics\ILevelTextureStorage.h" />
+    <ClInclude Include="Graphics\IMeshStorage.h" />
     <ClInclude Include="Graphics\ITextureStorage.h" />
     <ClInclude Include="Graphics\LevelTextureStorage.h" />
+    <ClInclude Include="Graphics\MeshStorage.h" />
     <ClInclude Include="Graphics\SectorHighlight.h" />
     <ClInclude Include="Graphics\SelectionRenderer.h" />
     <ClInclude Include="Graphics\TextureStorage.h" />
