+++ trview.ui.render/MapRenderer.h
@@ -14,7 +14,7 @@
 #include <trview.graphics/Texture.h>
 #include <trview.common/Point.h>
 #include <trview.common/Size.h>
-#include "trview\Room.h"
+#include <trview.app/Elements/Room.h>
 #include <trview.graphics/RenderTarget.h>
 #include <trview.common/Event.h>
 #include <trview.graphics/FontFactory.h>
